+++ examples/feet_to_meters.jl
@@ -11,28 +11,28 @@ calc_callback = TclTk.Callback(calc_function)
 
 interp = tk_start()
 
-top = TkToplevel(interp)
+top = Toplevel(interp)
 wm.title(top, "Feet to Meters")
 
-frame = TtkFrame(top, padding=(3, 3, 12, 12))
+frame = Frame(top, padding=(3, 3, 12, 12))
 frame.grid(column=0, row=0, sticky="nwes")
 
-feet = TtkEntry(frame, width=7, textvariable="feet")
+feet = Entry(frame, width=7, textvariable="feet")
 feet.grid(column=2, row=1, sticky="we")
 
-meters = TtkLabel(frame, textvariable="meters")
+meters = Label(frame, textvariable="meters")
 meters.grid(column=2, row=2, sticky="we")
 
-calc = TtkButton(frame, text="Calculate", command=calc_callback.name)
+calc = Button(frame, text="Calculate", command=calc_callback.name)
 calc.grid(column=3, row=3, sticky="w")
 
-flbl = TtkLabel(frame, text="feet")
+flbl = Label(frame, text="feet")
 flbl.grid(column=3, row=1, sticky="w")
 
-islbl = TtkLabel(frame, text="is equivalent to")
+islbl = Label(frame, text="is equivalent to")
 islbl.grid(column=1, row=2, sticky="e")
 
-mlbl = TtkLabel(frame, text="meters")
+mlbl = Label(frame, text="meters")
 mlbl.grid(column=3, row=2, sticky="w")
 
 interp(:grid, :columnconfigure, top, 0, weight=1)